moved optimizer-dependent data-model fields, and validators
