+++ bofire/data_models/strategies/api.py
@@ -19,12 +19,12 @@ from bofire.data_models.strategies.fractional_factorial import (
 from bofire.data_models.strategies.meta_strategy_type import MetaStrategy
 from bofire.data_models.strategies.predictives.acqf_optimization import (
     AcquisitionOptimizer,
-    BotorchOptimizer,
+    BotorchOptimizer, LSRBO,
 )
 from bofire.data_models.strategies.predictives.active_learning import (
     ActiveLearningStrategy,
 )
-from bofire.data_models.strategies.predictives.botorch import LSRBO, BotorchStrategy
+from bofire.data_models.strategies.predictives.botorch import BotorchStrategy
 from bofire.data_models.strategies.predictives.enting import EntingStrategy
 from bofire.data_models.strategies.predictives.mobo import MoboStrategy
 from bofire.data_models.strategies.predictives.multi_fidelity import (
