+++ bofire/strategies/predictives/acqf_optimization.py
@@ -60,3 +60,5 @@ class BotorchOptimizer(AcquisitionOptimizer):
     ) -> Tuple[torch.Tensor, torch.Tensor]:
         # this is the implementation of the optimizer, here goes _optimize_acqf_continuous
         pass
+
+
