+++ tests/bofire/data_models/specs/local_search_config.py
@@ -1,4 +1,4 @@
-from bofire.data_models.strategies.api import LSRBO
+from bofire.data_models.strategies.predictives.acqf_optimization import LSRBO
 from tests.bofire.data_models.specs.specs import Specs
