+++ tests/bofire/data_models/specs/strategies.py
@@ -1,4 +1,5 @@
 import bofire.data_models.strategies.api as strategies
+import bofire.data_models.strategies.predictives.acqf_optimization
 from bofire.data_models.acquisition_functions.api import (
     qEI,
     qLogNEHVI,
@@ -537,7 +538,7 @@ specs.add_invalid(
                 ],
             ),
         ).model_dump(),
-        "local_search_config": strategies.LSRBO(),
+        "local_search_config": bofire.data_models.strategies.predictives.acqf_optimization.LSRBO(),
     },
     error=ValueError,
     message="LSR-BO only supported for linear constraints.",
