+++ tests/bofire/data_models/strategies/predictives/test_botorch.py
@@ -2,7 +2,8 @@ import pytest
 
 from bofire.data_models.domain.api import Domain
 from bofire.data_models.features.api import ContinuousInput, ContinuousOutput
-from bofire.data_models.strategies.api import LSRBO, SoboStrategy
+from bofire.data_models.strategies.api import SoboStrategy
+from bofire.data_models.strategies.predictives.acqf_optimization import LSRBO
 
 
 @pytest.mark.parametrize(
