+++ tests/bofire/strategies/test_sobo.py
@@ -41,7 +41,7 @@ from bofire.data_models.objectives.api import (
     MaximizeObjective,
     MaximizeSigmoidObjective,
 )
-from bofire.data_models.strategies.api import LSRBO
+from bofire.data_models.strategies.predictives.acqf_optimization import LSRBO
 from bofire.data_models.strategies.api import RandomStrategy as RandomStrategyDataModel
 from bofire.data_models.unions import to_list
 from bofire.strategies.api import CustomSoboStrategy, RandomStrategy, SoboStrategy
